Ryddet i norg2.
