+++ src/main/kotlin/no/nav/tms/personopplysninger/api/Application.kt
@@ -41,6 +41,7 @@ fun main() {
             }
         },
         module = {
+            rootPath = "tms-personopplysninger-api"
             mainModule(personaliaService, httpClient)
         }
     ).start(wait = true)
