--- src/main/kotlin/no/nav/tms/personopplysninger/api/personalia/norg2/dto/Aapningstider.kt
@@ -1,11 +1,0 @@
-package no.nav.tms.personopplysninger.api.personalia.norg2.dto
-
-data class Aapningstider(
-    val dag: String?,
-    val dato: String?,
-    val fra: String?,
-    val til: String?,
-    val kommentar: String?,
-    val stengt: String?,
-    val kunTimeavtale: String?,
-)
