--- src/main/kotlin/no/nav/tms/personopplysninger/api/personalia/norg2/dto/Brukerkontakt.kt
@@ -1,5 +1,0 @@
-package no.nav.tms.personopplysninger.api.personalia.norg2.dto
-
-data class Brukerkontakt(
-    val publikumsmottak: List<Publikumsmottak>
-)
