--- src/main/kotlin/no/nav/tms/personopplysninger/api/personalia/norg2/dto/Norg2Enhet.kt
@@ -1,5 +1,0 @@
-package no.nav.tms.personopplysninger.api.personalia.norg2.dto
-
-data class Norg2Enhet(
-    val enhetNr: String = "",
-)
