--- src/main/kotlin/no/nav/tms/personopplysninger/api/personalia/norg2/dto/Norg2EnhetKontaktinfo.kt
@@ -1,6 +1,0 @@
-package no.nav.tms.personopplysninger.api.personalia.norg2.dto
-
-data class Norg2EnhetKontaktinfo(
-    val navn: String,
-    val brukerkontakt: Brukerkontakt,
-)
