--- src/main/kotlin/no/nav/tms/personopplysninger/api/personalia/norg2/dto/Publikumsmottak.kt
@@ -1,8 +1,0 @@
-package no.nav.tms.personopplysninger.api.personalia.norg2.dto
-
-data class Publikumsmottak(
-    val besoeksadresse: Adresse?,
-    val aapningstider: List<Aapningstider>?,
-    val stedsbeskrivelse: String?,
-    val adkomstbeskrivelse: String?,
-)
